Read sayings from spread sheets.

Read sayings from spread sheets.

read.sinxs(file = NULL, sep = ",")

Arguments

file
a character string giving a sinx sayings database in csv format (in UTF-8 encoding). By default all csv files in the data directory of the sinx package are used.

Value

a data frame of sayings, each row contains: - quote: the quote, main part of the sayings, - author: the author of the quote, - context: the context in which it was quoted (if available, otherwise NA), - source : where it was quoted (if available, otherwise NA), - date: when it was quoted (if available, otherwise NA).

Examples

read.sinxs()

#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           quote
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            这个问题我很感兴趣，然而遗憾的是，这个问题我无法重现。\\n\\n能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好能提供 `session_info()` 信息。\\n\\n提问题的方式详见[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)。
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Sweave 好像就是一边编译 LaTeX 的时候就可以一边执行程序源代码并即时输出结果，我也没正式用过……
#> 3 节约别人的时间，就是节约自己的时间。\\n举个例子，您是从医学的观点出发进行数据分析，对统计以及数理细节其实并不关心，属于“理解当然更好，但是没有也无所谓”的情况。而您真正关心的是，哪个r包，哪个算法，或者跟确切一点，哪个语句能给出您所期待的答案。\\n\\n为什么说会节约时间呢，很简单的一个问题，比如说对我来说，要复现你的问题，那首先得：\\n\\n1. 找一个p>n 的数据集\\n2. 研究一下glmnet怎么用\\n3. 再写一个3~10行的脚本跑一下看看结果\\n\\n虽然1，2，3都是很简单的很容易做到的工作，但是很麻烦，这又不是我的工作内容。虽然培训班网上教程都讲烂了，但是不好意思我没空去看那些。包的手册虽然也能看没问题，但是还是很麻烦。那既然那么麻烦，就还是算了吧。那把我知道的东西告诉你，能帮多少算多少，剩下的就只有祝你好运了。\\n但是如果贴上来的话，我就不用去做1，2，3了，那直接解决4还是一个比较有意思的活。虽然不能保证能解决，但是至少在我有空的时候（比如昨天）可以尝试尝试，如果做到了，那直接可以把代码给你，互相都能省很多时间。\\n\\n而且惰性是人类固有的东西，就比如打以上这段话的时间都够把网上培训班教程再看两遍再写一个脚本了，但是我还是懒得去做这些麻烦的事情而选择在这里慢慢打字，因为这样不费脑子，至少不费管理科的脑子。。。。。
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      奇怪了你的代码在我的机器上完全没有问题啊
#>     author                            context     source          date
#> 1 dapengde 耐心温馨欢欣地回答一个不规范的提问 d.cosx.org 2018 年 12 月
#> 2    yihui     <https://d.cosx.org/d/7673/15> d.cosx.org 2007 年 10 月
#> 3  wglaive    <https://d.cosx.org/d/420212/6> d.cosx.org 2018 年 10 月
#> 4   tctcab    <https://d.cosx.org/d/420198/2> d.cosx.org  2018 年 9 月