Find R sinx fortunes.

Find R sinx fortunes.

sinx(which = NULL, sinxs.data = NULL, fixed = TRUE, showMatches = FALSE,
    author = character(), ...)

Arguments

which
an integer specifying the row number of `sinxs.data`. Alternatively `which`` can be a character and `grep`` is used to try to find a suitable row.

sinxs.data
data frame containing a fortune in each row. By default the data from the 'sinx' package are used.

fixed
logical passed to `grep` if `which`` is a character, indicating if it should work (if `TRUE`, as by default) with a simple character string or (if `FALSE`) with regular expressions.

showMatches
if `which` is character, a logical indicating if `sinx()` should print all the row numbers of `sinxs.data` which match the `grep` search.

author
a character string to match (via `grep`) to the "authors" column of `sinxs.data`.

...
potential further arguments passed to `grep`.

Value

an object of class "sinx" which is a row from a data frame of fortunes (like those read in from read.sinxs).

Examples

sinx()

#> 
#> 这个问题我很感兴趣，然而遗憾的是，这个问题我无法重现。
#> 
#> 能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好能提供 `session_info()` 信息。
#> 
#> 提问题的方式详见[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)。
#> 
#> 
#>    -- dapengde (耐心温馨欢欣地回答一个不规范的提问)
#>       d.cosx.org (2018 年 12 月)
#> 

for (i in 1:4) sinx(i)

path_f <- system.file("fortunes/fortunes.csv", package = "fortunes")
path_s <- system.file("sinxs/sinxs.csv", package = "sinx")
ftns <- sinx::read.sinxs(c(path_f, path_s), sep = c(";", ","))
sinx::sinx(sinxs.data = ftns)

#> 
#> I think, therefore I R.
#> 
#> 
#>    -- William B. King (in his R tutorials)
#>       http://ww2.coastal.edu/kingw/statistics/R-tutorials/ (July 2010)
#>