Sling messages and warnings with flair

Sling messages and warnings with flair

tanx(which = NULL, sinxs.data = NULL, what = "Hello world!", by = "random",
    copy = FALSE, type = "message", what_color = "steelblue3", by_color = NULL,
    length = 18, ...)

Arguments

which
An integer specifying the row number of sinxs.data. Alternatively which can be a character and grep is used to try to find a suitable row.

what
(character) What do you want to say? See details.

by
(character) Type of thing, one of cow, chicken, poop, cat, facecat, bigcat, longcat, shortcat, behindcat, longtailcat, anxiouscat, grumpycat, smallcat, ant, pumpkin, ghost, spider, rabbit, pig, snowman, frog, hypnotoad, signbunny, stretchycat, fish, trilobite, shark, buffalo, clippy, mushroom, monkey, egret, or rms for Richard Stallman. Alternatively, use "random" to have your message spoken by a random character. We use match.arg internally, so you can use unique parts of words that don't conflict with others, like "g" for "ghost" because there's no other animal that starts with "g".

type
(character) One of message (default), warning, or string (returns string). If multiple colors are supplied to what_color or by_color, type cannot be warning. (This is a limitation of the multicolor packcage :/.)

what_color
(character or crayon function) One or more crayon-suported text color(s) or crayon style function to color what. You might try colors() or ?rgb for ideas. Use "rainbow" for c("red", "orange", "yellow", "green", "blue", "purple").

by_color
(character or crayon function) One or more crayon-suported text color(s) or crayon style function to color who. Use "rainbow" for c("red", "orange", "yellow", "green", "blue", "purple").

length
(integer) Length of longcat. Ignored if other animals used.

...
Further args passed on to sinx

Details

You can put in any phrase you like, OR you can type in one of a few special phrases that do particular things. They are:

catfact A random cat fact from https://catfact.ninja

fortune A random quote from an R coder, from fortunes library

time Print the current time

rms Prints a random 'fact' about Richard Stallman from the rmsfact package. Best paired with by = "rms".

Note that if you choose by='hypnotoad' the quote is forced to be, as you could imagine, 'All Glory to the HYPNO TOAD!'. For reference see http://knowyourmeme.com/memes/hypnotoad

Signbunny: It's not for sure known who invented signbunny, but this article http://www.vox.com/2014/9/18/6331753/sign-bunny-meme-explained thinks they found the first use in this tweet: https://twitter.com/wei_bluebear/status/329101645780770817

Trilobite: from http://www.retrojunkie.com/asciiart/animals/dinos.htm (site down though)

Note to Windows users: there are some animals (shortcat, longcat, fish, signbunny, stretchycat, anxiouscat, longtailcat, grumpycat, mushroom) that are not available because they use non-ASCII characters that don't display properly in R on Windows.

Examples

tanx()

#> 
#>     
#>  -------------- 
#> 这个问题我很感兴趣，然而遗憾的是，这个问题我无法重现。
#> 
#> 能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好能提供 `session_info()` 信息。
#> 
#> 提问题的方式详见[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)。
#> 
#> --- dapengde (耐心温馨欢欣地回答一个不规范的提问), d.cosx.org, 2018 年 12 月 
#>  --------------
#>     \
#>       \
#>         \
#>               /""-._
#>               .       '-,
#>                :          '',
#>                 ;      *     '.
#>                  ' *         () '.
#>                    \               \
#>                     \      _.---.._ '.
#>                     :  .' _.--''-''  \ ,'
#>         .._           '/.'             . ;
#>         ; `-.          ,                \'
#>          ;   `,         ;              ._\
#>           ;    \     _,-'                ''--._
#>           :    \_,-'                          '-._
#>           \ ,-'                       .          '-._
#>           .'         __.-'';            \...,__       '.
#>         .'      _,-'        \              \   ''--.,__  '\
#>         /    _,--' ;         \              ;           \^.}
#>         ;_,-' )     \  )\      )            ;
#>              /       \/  \_.,-'             ;
#>             /                              ;
#>          ,-'  _,-'''-.    ,-.,            ;      PFA
#>       ,-' _.-'        \  /    |/'-._...--'
#>      :--``             )/
#>   '
#>   

for (i in 1:4) tanx(i)

#> 
#> 
#>  ----- 
#> 这个问题我很感兴趣，然而遗憾的是，这个问题我无法重现。
#> 
#> 能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好能提供 `session_info()` 信息。
#> 
#> 提问题的方式详见[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)。
#> 
#> --- dapengde (耐心温馨欢欣地回答一个不规范的提问), d.cosx.org, 2018 年 12 月 
#>  ------ 
#>     \   
#>      \
#>                     @@@@@@ @
#>                   @@@@     @@
#>                  @@@@ =   =  @@ 
#>                 @@@ @ _   _   @@ 
#>                  @@@ @(0)|(0)  @@ 
#>                 @@@@   ~ | ~   @@
#>                 @@@ @  (o1o)    @@
#>                @@@    #######    @
#>                @@@   ##{+++}##   @@
#>               @@@@@ ## ##### ## @@@@
#>               @@@@@#############@@@@
#>              @@@@@@@###########@@@@@@
#>             @@@@@@@#############@@@@@
#>             @@@@@@@### ## ### ###@@@@
#>              @ @  @              @  @
#>                @                    @

#> 
#>  -------------- 
#> Sweave 好像就是一边编译 LaTeX 的时候就可以一边执行程序源代码并即时输出结果，我也没正式用过……
#> 
#> --- yihui (<https://d.cosx.org/d/7673/15>), d.cosx.org, 2007 年 10 月 
#>  --------------
#>     \
#>       \
#>          /\ /\
#>          (O o)
#>         =(:^:)=
#>            U      [nosig]
#>   

#> 
#>  -------------- 
#> 节约别人的时间，就是节约自己的时间。
#> 举个例子，您是从医学的观点出发进行数据分析，对统计以及数理细节其实并不关心，属于“理解当然更好，但是没有也无所谓”的情况。而您真正关心的是，哪个r包，哪个算法，或者跟确切一点，哪个语句能给出您所期待的答案。
#> 
#> 为什么说会节约时间呢，很简单的一个问题，比如说对我来说，要复现你的问题，那首先得：
#> 
#> 1. 找一个p>n 的数据集
#> 2. 研究一下glmnet怎么用
#> 3. 再写一个3~10行的脚本跑一下看看结果
#> 
#> 虽然1，2，3都是很简单的很容易做到的工作，但是很麻烦，这又不是我的工作内容。虽然培训班网上教程都讲烂了，但是不好意思我没空去看那些。包的手册虽然也能看没问题，但是还是很麻烦。那既然那么麻烦，就还是算了吧。那把我知道的东西告诉你，能帮多少算多少，剩下的就只有祝你好运了。
#> 但是如果贴上来的话，我就不用去做1，2，3了，那直接解决4还是一个比较有意思的活。虽然不能保证能解决，但是至少在我有空的时候（比如昨天）可以尝试尝试，如果做到了，那直接可以把代码给你，互相都能省很多时间。
#> 
#> 而且惰性是人类固有的东西，就比如打以上这段话的时间都够把网上培训班教程再看两遍再写一个脚本了，但是我还是懒得去做这些麻烦的事情而选择在这里慢慢打字，因为这样不费脑子，至少不费管理科的脑子。。。。。
#> 
#> --- wglaive ( <https://d.cosx.org/d/420212/6>), d.cosx.org, 2018 年 10 月 
#>  --------------
#>       \
#>         \
#>           \
#>             |\___/|
#>             )     (
#>            =\     /=
#>              )===(
#>             /     \
#>             |     |
#>            /       \
#>            \       /
#>       jgs   \__  _/
#>               ( (
#>                ) )
#>               (_(
#>   

#> 
#> 
#>  ----- 
#> 奇怪了你的代码在我的机器上完全没有问题啊
#> 
#> --- tctcab (<https://d.cosx.org/d/420198/2>), d.cosx.org, 2018 年 9 月 
#>  ------ 
#>     \   
#>      \
#>               |
#>               |
#>               |
#>              __
#>           | /  \ |
#>          \_\\  //_/
#>           .'/()\'.
#>            \\  //  [nosig]
#>   

path_f <- system.file("fortunes/fortunes.csv", package = "fortunes")
path_s <- system.file("sinxs/sinxs.csv", package = "sinx")
ftns <- sinx::read.sinxs(c(path_f, path_s), sep = c(";", ","))
sinx::tanx(sinxs.data = ftns)

#> 
#> 
#>  ------------- 
#> Kevin Wright: Doug already mentioned the limitations of open source optimization code. Open source has benefits, but simply being open doesn't necessarily mean better.<x><x>Joshua Wiley: Actually, a series of studies I have conducted including theoretical proofs and empirical evidence proves open source to be unequivocally better. If you are interested, you can pay me to use the assertion, but the supporting details are restricted so you must accept that my internal checks and quality control are sufficient and the findings are correct and accurate. Details follow.<x><x>License and Use of I. Ben Fooled's findings:<x><x>"Open source is unequivocally better" is $10 for personal use or $20 for each public use of the assertion. It is also possible to get a yearly license, but note that this does not include updates. Support is extra. If you wish to use the assertions on a multicore system, there is an additional $5 per core per assertion fee. A student version is available for pedagogical purposes only: "pn src s nqvcll bttr". If you need more than 20 letters, please upgrade to the full version.<x><x>This is the best thing since bread, but don't take my word for it. Here are some trite testimonials putatively from customers:<x><x>"Using I. Ben Fooled's assertion has really revolutionized my work. I love it. I use it every day" ~ Jane, Data Expert<x><x>"I lost 130 lbs on this system", 'wait, this isn't a weightloss testimonial? Sorry!' "My work is easily twice as efficient since I switched to I. Ben Fooled's product." ~ John, Entrepeneur
#> 
#> --- Kevin Wright and Joshua Wiley (NA), R-SIG-Mixed-Models, October 2011 
#>  -------------- 
#>               \   
#>                \  
#>                 \
#>         __.--'\     \.__./     /'--.__
#>     _.-'       '.__.'    '.__.'       '-._
#>   .'                                      '.
#>  /                                          \
#> |                                            |
#> |                                            |
#>  \         .---.              .---.         /
#>   '._    .'     '.''.    .''.'     '.    _.'
#>      '-./            \  /           \.-'
#>                       ''mrf