Sino Xmen IN the R community.

Sino Xmen IN the R community.

sinx(which = NULL, sinxs.data = NULL, fixed = TRUE, showMatches = FALSE,
    author = character(), ...)

Arguments

which
an integer specifying the row number of `sinxs.data`. Alternatively `which`` can be a character and `grep`` is used to try to find a suitable row.

sinxs.data
data frame containing a saying in each row. By default the data from the 'sinx' package are used.

fixed
logical passed to `grep` if `which`` is a character, indicating if it should work (if `TRUE`, as by default) with a simple character string or (if `FALSE`) with regular expressions.

showMatches
if `which` is character, a logical indicating if `sinx()` should print all the row numbers of `sinxs.data` which match the `grep` search.

author
a character string to match (via `grep`) to the "authors" column of `sinxs.data`.

...
potential further arguments passed to `grep`.

Value

an object of class "sinx" which is a row from a data frame of sayings (like those read in from read.sinxs).

Examples

sinx()

#> 
#> 这个问题我很感兴趣。能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好附带 `session_info()` 信息。
#> 
#> 提问题的方式可以参考“[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)”。
#> 
#> 
#>    -- dapengde (<https://d.cosx.org/d/420469/2>)
#>       d.cosx.org (2018-12)
#> 

for (i in 1:4) print(sinx(i))

#> 
#> 这个问题我很感兴趣。能否提供一个可以重现问题的数据和代码示例，让我直接拷贝粘贴就能运行？最好附带 `session_info()` 信息。
#> 
#> 提问题的方式可以参考“[新手须知 1.2 正确的发帖姿势](https://d.cosx.org/d/1553)”。
#> 
#> 
#>    -- dapengde (<https://d.cosx.org/d/420469/2>)
#>       d.cosx.org (2018-12)
#> 
#> 
#> Sweave 好像就是一边编译 LaTeX 的时候就可以一边执行程序源代码并即时输出结果，我也没正式用过……
#> 
#> 
#>    -- yihui (<https://d.cosx.org/d/7673/15>)
#>       d.cosx.org (2007-10)
#> 
#> 
#> 节约别人的时间，就是节约自己的时间。
#> 举个例子，您是从医学的观点出发进行数据分析，对统计以及数理细节其实并不关心，属于“理解当然更好，但是没有也无所谓”的情况。而您真正关心的是，哪个r包，哪个算法，或者跟确切一点，哪个语句能给出您所期待的答案。
#> 
#> 为什么说会节约时间呢，很简单的一个问题，比如说对我来说，要复现你的问题，那首先得：
#> 
#> 1. 找一个p>n 的数据集
#> 2. 研究一下glmnet怎么用
#> 3. 再写一个3~10行的脚本跑一下看看结果
#> 
#> 虽然1，2，3都是很简单的很容易做到的工作，但是很麻烦，这又不是我的工作内容。虽然培训班网上教程都讲烂了，但是不好意思我没空去看那些。包的手册虽然也能看没问题，但是还是很麻烦。那既然那么麻烦，就还是算了吧。那把我知道的东西告诉你，能帮多少算多少，剩下的就只有祝你好运了。
#> 但是如果贴上来的话，我就不用去做1，2，3了，那直接解决4还是一个比较有意思的活。虽然不能保证能解决，但是至少在我有空的时候（比如昨天）可以尝试尝试，如果做到了，那直接可以把代码给你，互相都能省很多时间。
#> 
#> 而且惰性是人类固有的东西，就比如打以上这段话的时间都够把网上培训班教程再看两遍再写一个脚本了，但是我还是懒得去做这些麻烦的事情而选择在这里慢慢打字，因为这样不费脑子，至少不费管理科的脑子。。。。。
#> 
#> 
#>    -- wglaive (<https://d.cosx.org/d/420212/6>)
#>       d.cosx.org (2018-10)
#> 
#> 
#> 奇怪了你的代码在我的机器上完全没有问题啊
#> 
#> 
#>    -- tctcab (<https://d.cosx.org/d/420198/2>)
#>       d.cosx.org (2018-09)
#> 

path_f <- system.file("fortunes/fortunes.csv", package = "fortunes")
path_s <- system.file("sinxs/sinxs.csv", package = "sinx")
ftns <- sinx::read.sinxs(c(path_f, path_s), sep = c(";", ","))
sinx::sinx(sinxs.data = ftns)

#> 
#> Jeremy Koster: My students were looking at the estimated varying intercepts for each higher-level group (or the "BLUP's", as some people seem to call them).<x>Douglas Bates: As Alan James once said, "these values are just like the BLUPs - Best Linear Unbiased Predictors - except that they aren't linear and they aren't unbiased and there is no clear sense in which they are "best", but other than that ..."
#> 
#> 
#>    -- Jeremy Koster and Douglas Bates
#>       R-SIG-Mixed-Models (July 2012)
#> 

jinyong <- read.sinxs(lib = "jinyong")
sinx(sinxs.data = jinyong)

#> 
#> 他强由他强，清风拂山冈；他横任他横，明月照大江。
#> 
#> 
#>    -- 金庸 (55、 他强由他强，清风拂山冈；他横任他横，明月照大江。 ——金庸)
#>       
#> 

libs <- read.sinxs(lib = c("tangshi", "songshi", "chinese", "yangsheng", "english",
    "jinyong"))
sinx(sinxs.data = libs)

#> 
#> 人没钱就像弓没箭
#> 
#> 
#>    -- 托·富勒
#>       
#>